TCP and database

+++ schema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS ecFlex_data(
     ID INTEGER,
-    Timer INTEGER,
-    Temperature INTEGER,
+    Timer FLOAT,
+    Temperature FLOAT,
     Glucose_concentration FLOAT
 );
